+++ max/exports/validatePhoneNumberLength.js
@@ -1,0 +1,6 @@
+import { withMetadata } from '../metadata'
+import { validatePhoneNumberLength as _validatePhoneNumberLength } from '../../core/index'
+
+export function validatePhoneNumberLength() {
+	return withMetadata(_validatePhoneNumberLength, arguments)
+}