+++ source/AsYouType.test.js
@@ -362,10 +362,12 @@ describe('AsYouType', () => {
 		// formatter.input('+1 1 877 215 5230').should.equal('+1 1 8772155230')
 		formatter.getNationalNumber().should.equal('8772155230')
 
-		formatter = new AsYouType()
-		formatter.input('+78800555353').should.equal('+7 880 055 53 53')
-		formatter.input('5').should.equal('+7 8 800 555 35 35')
-		formatter.getNationalNumber().should.equal('8005553535')
+		// They've added another number format that has `8` leading digit
+		// and 14 digits. Maybe it's something related to Kazakhstan.
+		// formatter = new AsYouType()
+		// formatter.input('+78800555353').should.equal('+7 880 055 53 53')
+		// formatter.input('5').should.equal('+7 8 800 555 35 35')
+		// formatter.getNationalNumber().should.equal('8005553535')
 	})
 
 	it('should return a partial template for current value', () => {
@@ -923,10 +925,10 @@ describe('AsYouType', () => {
 
 	it('shouldn\'t choose a format when there\'re too many digits for any of them', () => {
 		const formatter = new AsYouType('RU')
-		formatter.input('88005553535')
+		formatter.input('89991112233')
 		formatter.formatter.chosenFormat.format().should.equal('$1 $2-$3-$4')
 		formatter.reset()
-		formatter.input('880055535355')
+		formatter.input('899911122334')
 		expect(formatter.formatter.chosenFormat).to.be.undefined
 	})